💬 Address PR comments 🛠  Fix mypy issues
